requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0
python-decouple>=3.8
django-ckeditor>=6.7.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0

